+++ packages/core-mobile/app/new/features/swap/hooks/useQuickSwaps.ts
@@ -1,6 +1,4 @@
-import { useCallback } from 'react'
 import { useSelector } from 'react-redux'
-import type Big from 'big.js'
 import { useActiveWallet } from 'common/hooks/useActiveWallet'
 import { selectActiveNetwork } from 'store/network/slice'
 import { selectIsQuickSwapsAvailable } from 'store/posthog'
@@ -15,7 +13,6 @@ import type {
 } from 'store/settings/advanced/types'
 import { WalletType } from 'services/wallet/types'
 import { NetworkVMType } from '@avalabs/core-chains-sdk'
-import { isAmountOverLimit } from '../utils/quickSwapsLimits'
 
 // Allowlist (not denylist): future wallet types fail-safe — they must be
 // explicitly added here to become eligible.
@@ -39,7 +36,6 @@ type UseQuickSwapsResult = {
   isEnabled: boolean
   feeSetting: QuickSwapFeeLevel
   maxBuy: QuickSwapMaxBuy
-  isAmountOverLimit: (amountUsd: Big | undefined) => boolean
 }
 
 export const useQuickSwaps = (): UseQuickSwapsResult => {
@@ -57,11 +53,6 @@ export const useQuickSwaps = (): UseQuickSwapsResult => {
   const chainAllowed = activeNetwork?.vmName === NetworkVMType.EVM
   const isAvailable = flagOn && walletAllowed && chainAllowed
 
-  const overLimit = useCallback(
-    (amountUsd: Big | undefined) => isAmountOverLimit(amountUsd, maxBuy),
-    [maxBuy]
-  )
-
   return {
     flagOn,
     walletAllowed,
@@ -69,7 +60,6 @@ export const useQuickSwaps = (): UseQuickSwapsResult => {
     isAvailable,
     isEnabled: isAvailable && rawIsEnabled,
     feeSetting,
-    maxBuy,
-    isAmountOverLimit: overLimit
+    maxBuy
   }
 }
