--- packages/core-mobile/app/new/features/swap/utils/quickSwapsLimits.ts
@@ -1,11 +1,0 @@
-import Big from 'big.js'
-import { QuickSwapMaxBuy } from 'store/settings/advanced/types'
-
-export const isAmountOverLimit = (
-  amountUsd: Big | undefined,
-  maxBuy: QuickSwapMaxBuy
-): boolean => {
-  if (maxBuy === 'unlimited') return false
-  if (amountUsd === undefined) return true
-  return amountUsd.gt(new Big(maxBuy))
-}
