+++ packages/core-mobile/app/vmModule/ApprovalController/validators/shared.ts
@@ -1,0 +1,223 @@
+import { AlertType, type RpcRequest } from '@avalabs/vm-module-types'
+import { isInAppRequest } from 'store/rpc/utils/isInAppRequest'
+import { WalletType } from 'services/wallet/types'
+import { RequestContext, type SwapAutoApproveContext } from 'store/rpc/types'
+import {
+  validateSwapAmounts,
+  type BalanceChangeData,
+  type ValidationFailReason
+} from 'features/swap/utils/swapValidation'
+import AnalyticsService from 'services/analytics/AnalyticsService'
+import Logger from 'utils/Logger'
+import type { ValidationResult } from './types'
+
+// Allowlist (not denylist): future wallet types fail-safe — they must be
+// explicitly added here to become eligible. Also fails closed when
+// walletType is missing from context (denylist would fail open).
+const SOFTWARE_WALLET_TYPES: ReadonlySet<WalletType> = new Set([
+  WalletType.MNEMONIC,
+  WalletType.SEEDLESS,
+  WalletType.PRIVATE_KEY
+])
+
+type BypassFallbackReason =
+  | 'context_missing'
+  | 'tx_flagged_warning'
+  | 'tx_flagged_malicious'
+  | 'unknown'
+  | ValidationFailReason
+
+const readAutoApproveContext = (
+  request: RpcRequest
+): SwapAutoApproveContext | undefined => {
+  const ctx = request.context as Record<string, unknown> | undefined
+  if (!ctx) return undefined
+  const value = ctx[RequestContext.SWAP_AUTO_APPROVE]
+  if (
+    value !== null &&
+    typeof value === 'object' &&
+    typeof (value as { autoApprove?: unknown }).autoApprove === 'boolean'
+  ) {
+    return value as SwapAutoApproveContext
+  }
+  return undefined
+}
+
+const readWalletType = (request: RpcRequest): WalletType | undefined => {
+  const ctx = request.context as Record<string, unknown> | undefined
+  const wt = ctx?.walletType
+  if (
+    typeof wt === 'string' &&
+    Object.values(WalletType).includes(wt as WalletType)
+  ) {
+    return wt as WalletType
+  }
+  return undefined
+}
+
+// Trust boundary: isInAppRequest gate prevents external dApps from
+// triggering the bypass even if their context shape matches. Method
+// check stays with each validator since it differs between single-tx
+// and batch.
+export const isBypassEligible = (request: RpcRequest): boolean => {
+  if (!isInAppRequest(request)) return false
+  const ctx = readAutoApproveContext(request)
+  if (!ctx?.autoApprove) return false
+  const walletType = readWalletType(request)
+  return walletType !== undefined && SOFTWARE_WALLET_TYPES.has(walletType)
+}
+
+const fallback = (
+  reason: string,
+  code: BypassFallbackReason
+): ValidationResult => ({
+  isValid: false,
+  requiresManualApproval: true,
+  reason,
+  code: code as ValidationFailReason
+})
+
+const hardReject = (
+  reason: string,
+  code: BypassFallbackReason
+): ValidationResult => ({
+  isValid: false,
+  requiresManualApproval: false,
+  reason,
+  code: code as ValidationFailReason
+})
+
+// Minimal structural type — single-tx and batch displayData both expose
+// these three fields with identical shapes.
+type DisplayDataLike = {
+  alert?: {
+    type: AlertType
+    details?: { title?: string; description?: string }
+  }
+  isSimulationSuccessful?: boolean
+  balanceChange?: unknown
+}
+
+// Both Danger and Warning fall through to the manual modal — matches
+// `core-extension`. The modal renders the alert banner inline (with
+// Blockaid's "Scam transaction" / "Proceed Anyway" buttons) so the
+// user sees the warning and chooses. Hard-rejecting would surface a
+// confusing error toast with no path to proceed.
+const interpretAlert = (
+  alert: DisplayDataLike['alert']
+): ValidationResult | null => {
+  if (!alert) return null
+  if (alert.type === AlertType.DANGER) {
+    return fallback(
+      alert.details?.title || 'Transaction flagged as malicious',
+      'tx_flagged_malicious'
+    )
+  }
+  if (alert.type === AlertType.WARNING) {
+    return fallback(
+      alert.details?.title || 'Transaction safety check returned Warning',
+      'tx_flagged_warning'
+    )
+  }
+  return null
+}
+
+const runSwapValidation = async (params: {
+  request: RpcRequest
+  displayData: DisplayDataLike
+}): Promise<ValidationResult> => {
+  const { request, displayData } = params
+  const ctx = readAutoApproveContext(request)
+  if (!ctx) {
+    return fallback('Quick Swaps context not provided', 'context_missing')
+  }
+
+  const alertResult = interpretAlert(displayData.alert)
+  if (alertResult) return alertResult
+
+  if (displayData.isSimulationSuccessful === false) {
+    return hardReject(
+      'Transaction simulation failed - cannot safely auto-approve swap',
+      'simulation_failed'
+    )
+  }
+
+  return validateSwapAmounts({
+    displayData: {
+      isSimulationSuccessful: displayData.isSimulationSuccessful,
+      // SDK BalanceChange and local BalanceChangeData are structurally
+      // identical but nominally distinct — cast at the boundary.
+      balanceChange: displayData.balanceChange as BalanceChangeData | undefined
+    },
+    context: {
+      srcTokenAddress: ctx.srcTokenAddress,
+      destTokenAddress: ctx.destTokenAddress,
+      isSrcTokenNative: ctx.isSrcTokenNative ?? false,
+      isDestTokenNative: ctx.isDestTokenNative ?? false,
+      slippage: ctx.slippage,
+      minAmountOut: ctx.minAmountOut,
+      amountIn: ctx.amountIn,
+      maxBuy: ctx.maxBuy,
+      isSwapFeesEnabled: ctx.isSwapFeesEnabled ?? false
+    }
+  })
+}
+
+const captureFired = (
+  request: RpcRequest,
+  ctx: SwapAutoApproveContext
+): void => {
+  AnalyticsService.capture('QuickSwapsBypassFired', {
+    caip2SourceChainId: request.chainId,
+    maxBuy: ctx.maxBuy ?? 'unlimited'
+  })
+}
+
+const captureFellBack = (
+  request: RpcRequest,
+  reason: BypassFallbackReason,
+  requiresManualApproval: boolean
+): void => {
+  AnalyticsService.capture('QuickSwapsBypassFellBack', {
+    caip2SourceChainId: request.chainId,
+    reason,
+    requiresManualApproval
+  })
+}
+
+// Wraps validation with try/catch and analytics. Single-tx and batch
+// validators share this whole flow — only canHandle (method check)
+// differs.
+export const runValidateAndCapture = async (params: {
+  request: RpcRequest
+  displayData: DisplayDataLike
+  loggerTag: string
+}): Promise<ValidationResult> => {
+  let result: ValidationResult
+  try {
+    result = await runSwapValidation(params)
+  } catch (err) {
+    Logger.warn(`${params.loggerTag} validation failed; falling back`, err)
+    result = fallback('Transaction safety check failed', 'unknown')
+  }
+
+  if (result.isValid) {
+    const ctx = readAutoApproveContext(params.request)
+    if (ctx) captureFired(params.request, ctx)
+  } else {
+    // Logger.info, not warn — warn wraps in console.groupCollapsed
+    // which Metro swallows.
+    Logger.info(`${params.loggerTag} fell back to manual`, {
+      code: result.code,
+      reason: result.reason,
+      requiresManualApproval: result.requiresManualApproval
+    })
+    captureFellBack(
+      params.request,
+      (result.code as BypassFallbackReason | undefined) ?? 'unknown',
+      result.requiresManualApproval
+    )
+  }
+
+  return result
+}
